STORM-697: Fixed incorrect typing for offset

+++ external/storm-kafka/src/jvm/storm/kafka/KafkaUtils.java
@@ -211,7 +211,7 @@ public class KafkaUtils {
         return tups;
     }
     
-    public static Iterable<List<Object>> generateTuples(KafkaConfig kafkaConfig, Message msg, Partition partition, int offset) {
+    public static Iterable<List<Object>> generateTuples(KafkaConfig kafkaConfig, Message msg, Partition partition, long offset) {
         Iterable<List<Object>> tups;
         ByteBuffer payload = msg.payload();
         if (payload == null) {
