+++ external/storm-kafka/src/jvm/storm/kafka/StringMessageAndMetadataScheme.java
@@ -12,7 +12,7 @@ public class StringMessageAndMetadataScheme extends StringScheme implements Mess
     public static final String STRING_SCHEME_OFFSET = "offset";
 
     @Override
-    public List<Object> deserializeMessageWithMetadata(byte[] message, Partition partition, int offset) {
+    public List<Object> deserializeMessageWithMetadata(byte[] message, Partition partition, long offset) {
         String stringMessage = StringScheme.deserializeString(message);
         return new Values(stringMessage, partition.partition, offset);
     }
